Hoping that refactoring will fix breaking test on Travis
