+++ test/rest/histories.js
@@ -26,7 +26,7 @@ const historySchemas = require('./schemas/histories');
 const roomSchemas = require('./schemas/rooms');
 const errorSchemas = require('./schemas/errors');
 const testData = require('./helper/testData');
-const tokens = require('./helper/starter').tokens;
+const tokens = require('./0- starter').tokens;
 
 chai.should();
 chai.use(chaiHttp);
