+++ test/rest/messages.js
@@ -26,7 +26,7 @@ const messageSchemas = require('./schemas/messages');
 const errorSchemas = require('./schemas/errors');
 const roomSchemas = require('./schemas/rooms');
 const testData = require('./helper/testData');
-const tokens = require('./helper/starter').tokens;
+const tokens = require('./0- starter').tokens;
 
 chai.should();
 chai.use(chaiHttp);
