+++ test/rest/users.js
@@ -27,7 +27,7 @@ const authenticateSchemas = require('./schemas/authentications');
 const successSchemas = require('./schemas/successes');
 const errorSchemas = require('./schemas/errors');
 const testData = require('./helper/testData');
-const tokens = require('./helper/starter').tokens;
+const tokens = require('./0- starter').tokens;
 
 chai.should();
